+++ aws-lambda-function/main.tf
@@ -11,8 +11,7 @@ locals {
   }
 }
 
-
-resource "aws_lambda_function" "lambda" {
+resource aws_lambda_function lambda {
   s3_bucket = var.source_s3_bucket
   s3_key    = var.source_s3_key
 
@@ -53,7 +52,7 @@ resource "aws_lambda_function" "lambda" {
   tags = local.tags
 }
 
-data "aws_iam_policy_document" "lambda_role_policy" {
+data aws_iam_policy_document lambda_role_policy {
   statement {
     principals {
       type = "Service"
@@ -66,13 +65,11 @@ data "aws_iam_policy_document" "lambda_role_policy" {
   }
 }
 
-resource "aws_iam_role" "role" {
-  name = local.name
-  path = var.lambda_role_path
-
+resource aws_iam_role role {
+  name               = local.name
+  path               = var.lambda_role_path
   assume_role_policy = data.aws_iam_policy_document.lambda_role_policy.json
-
-  tags = local.tags
+  tags               = local.tags
 }
 
 resource "aws_cloudwatch_log_group" "log" {
@@ -84,14 +81,6 @@ data "aws_region" "current" {}
 data "aws_caller_identity" "current" {}
 
 # TODO scope this policy down
-#
-# I would love to use "${aws_cloudwatch_log_group.log.arn}", as the
-# resource here, but the provider returns an ARN that looks like:
-#   arn:aws:logs:us-west-2:123456789:log-group:/foo/bar:*
-# Unfortunately you need to use an ARN like:
-#   arn:aws:logs:us-west-2:123456789:log-group:/foo/bar
-# to match operations on the log group(like creating a new stream.) So instead we construct one
-# without the colon before the *, so that we can match both log groups and log streams.
 data "aws_iam_policy_document" "lambda_logging_policy" {
   statement {
     effect = "Allow"
@@ -104,7 +93,7 @@ data "aws_iam_policy_document" "lambda_logging_policy" {
     resources = [
       var.at_edge ?
       "*" :
-      "arn:aws:logs:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:log-group:${aws_cloudwatch_log_group.log.name}*",
+      "${aws_cloudwatch_log_group.log.arn}:*"
     ]
   }
 }
@@ -113,8 +102,7 @@ resource "aws_iam_policy" "lambda_logging" {
   name_prefix = "${local.name}-lambda-logging"
   path        = "/"
   description = "IAM policy for logging from the ${local.name} lambda."
-
-  policy = data.aws_iam_policy_document.lambda_logging_policy.json
+  policy      = data.aws_iam_policy_document.lambda_logging_policy.json
 }
 
 resource "aws_iam_role_policy_attachment" "lambda_logs" {